feat(backups): extract Alpine factory to static js module
Phase 2 of #827. Move the inline `<script>` body from
`backups_scripts.go` (Go-string raw literal) into a proper static JS
module at `static/js/admin/components/backups.js`, loaded via
`<script src="...">` at the top of the templ component.

The original IIFE relied on running AFTER the rest of the page body
was parsed (it queried `#schedule` / `#restore` for hash-scroll). Now
that the script loads synchronously at the top of the component
(required so `alpine:init` listeners register before Alpine fires),
the IIFE waits for `DOMContentLoaded` before scrolling.

The `alpine:init` listener that registers the `backups.new` keyboard
shortcut on `Alpine.store('shortcuts')` is unchanged.

Browser-validated against `/backups` (logged-in admin): page renders,
console silent, `Alpine.store('shortcuts').items` includes
`backups.new`, navigating to `/backups#schedule` scrolls the schedule
section into view.
